test(msvc): clear remaining /WX warnings (fdopen shim, more SOCKET/int64/_Atomic)
- unistd shim: fdopen()->_fdopen() (test_cfg C4996).
- test_net: (int) cast the two remaining socket() results (C4244).
- test_blocking: elapsed_ms is int64_t, not long (C4244 int64->long).
- test_coverage_pump: cast (void *)&counter at the 3 async call sites
  (C4090 _Atomic-qualifier drop).

+++ test/m2/test_net.c
@@ -104,7 +104,7 @@ test_tcp_knobs(const MunitParameter p[], void *d)
 	int v;
 	socklen_t l;
 	(void)p; (void)d;
-	fd = socket(AF_INET, SOCK_STREAM, 0);
+	fd = (int)socket(AF_INET, SOCK_STREAM, 0);
 	munit_assert_int(fd, >=, 0);
 	opts.nodelay = 1; opts.reuseaddr = 1; opts.keepalive = 1;
 	munit_assert_int(xtc_net_apply_tcp_opts(fd, &opts), ==, XTC_OK);
@@ -165,7 +165,7 @@ test_setnonblock(const MunitParameter p[], void *d)
 {
 	int fd;
 	(void)p; (void)d;
-	fd = socket(AF_INET, SOCK_STREAM, 0);
+	fd = (int)socket(AF_INET, SOCK_STREAM, 0);
 	munit_assert_int(fd, >=, 0);
 	munit_assert_int(xtc_net_setnonblock(fd), ==, XTC_OK);
 	close(fd);
